Redo insert-tag commands to use tag attribute and body definitions
This is a big commit. The defmacro/eval trick with
`bbcode-make-key-binding` is replaced with no eval, but unfortunately
a bigger macro `bbcode-define-insert-tag-commands`.

The old feature to use a prefix argument (C-u) for insert-tag commands
is gone. Instead, the user is always prompted for tag attributes in
the minibuffer (for tags that have attributes). There is one
minibuffer prompt per each attribute. The user can leave any attribute
blank to omit it.

Point is always put just before the closing tag. For multi-line tags,
we insert a blank line between the opening and closing tags and put
point there.